Bug fix for time rollover and reduced refresh rate. Code cleanup.
Found bug when using start time with non-zero minutes, rollover of hours
not occurring as should be so updated to use improved and simpler time
calculation.
Also reduced refresh rate which seems to improve the darkness of LED
segments that should be off.
Made separate function to display digits that could be reused for other
functionality.